+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 * Add expression for `(alias _name_)` in `(deps)`.
+* Add `(file_deps [...])` in `(deps)` for multiple dependencies.
 * Add `(with-stdin-from)` action.
 
 ## 0.1.0 (2022-10-14)